Skip Travis trying to create symlinks
Fails for unknown reasons, see build 1116 which has debug output.

+++ buildSrc/src/main/kotlin/configurations/FileUtils.kt
@@ -6,6 +6,11 @@ import java.nio.file.Files
 
 object FileUtils {
     fun Task.symlink(linkName: File, originFile: File) {
+        if (System.getenv("CI") != null) { // Travis can't do symlinks for unknown reasons
+            logger.warn("Detected CI environment, not creating any symlinks!")
+            return
+        }
+
         if (!linkName.exists() || linkName.delete()) {
             Files.createSymbolicLink(linkName.toPath(), originFile.toPath())
         } else {
